nultiples changes

+++ src/pt/isec/pa/javalife/model/data/Pedra.java
@@ -4,7 +4,7 @@ import pt.isec.pa.javalife.model.data.Inanimado;
 
 public final class Pedra extends Inanimado {
 
-    public Pedra(int cima, int esquerda, int baixo, int direita) {
+    public Pedra(double cima, double esquerda, double baixo, double direita) {
         super(new Area (cima, esquerda, baixo, direita));;
     }
 }
